requirements.txt
fastapi
uvicorn
playwright
openai
pandas
requests
beautifulsoup4
python-multipart
python-dotenv
websockets
pypdf
pillow
matplotlib
seaborn
scikit-learn
openai-whisper
